redo a structure of the project, added routing
