+++ client/src/components/assets/Banner/Banner.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import TabList from "../TabList/TabList";
+
+const Banner = () => {
+    return (
+        <TabList />
+    );
+};
+
+export default Banner;
