+++ client/src/components/assets/Body/Body.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import Banner from "../Banner/Banner";
+import Cadrlist from "../Cardlist/Cadrlist";
+
+const Body = () => {
+    return (
+        <div>
+            <Banner/>
+            <Cadrlist/>
+        </div>
+    );
+};
+
+export default Body;
