+++ client/src/components/assets/Cardlist/Cadrlist.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Cadrlist = () => {
+    return (
+        <div>
+            list
+        </div>
+    );
+};
+
+export default Cadrlist;
