+++ client/src/components/assets/Footer/Footer.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Footer = () => {
+    return (
+        <div>
+            СЮДИ КОД ГРИГОРІЯ
+        </div>
+    );
+};
+
+export default Footer;
