+++ client/src/components/assets/Header/Header.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Header = () => {
+    return (
+        <div>
+            СЮДИ КОД ГРИГОРІЯ
+        </div>
+    );
+};
+
+export default Header;
