+++ client/src/components/pages/NotFound.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const NotFound = () => {
+    return (
+        <div>
+            тут стилізація і кнопка на Ноme.
+            для переходу використати роутінг реакту
+        </div>
+    );
+};
+
+export default NotFound;
